Add files via upload
ddpm_flax_oxford102_end_to_end.ipynb is the colab that I almost got to work, but got stuck never finishing executing the training with the pretrained material. The other material in here is from when I was trying to get Imagen to work, which it didn't in Colab because there is not enough memory on the basic plan and didn't run at all in my local VSCode.